Issue #43: Split document definitions validator specs
